fasalrakshak/backend/ai_service/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
tensorflow-cpu==2.17.0
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
joblib==1.4.2
pillow==10.4.0
python-multipart==0.0.12
pydantic==2.9.2
